Fix test 7

+++ test/tests.cpp
@@ -46,7 +46,7 @@ TEST(MyStackTest, test6) {
 }
 
 TEST(MyStackTest, test7) {
-    MyStack<float> st(5);
+    MyStack<double> st(5);
     st.push(10.1);
     DOUBLE_EQ(10.1, st.get());
 }
